Improved surveypro utils
